new modularization, change date format by data-fns
