+++ index.html
@@ -1,15 +1,13 @@
 <!DOCTYPE html>
 <html>
+  <head>
+    <title>Проект "Комменты"</title>
+    <meta charset="utf-8" />
+    <link rel="stylesheet" href="styles.css" />
+  </head>
 
-<head>
-  <title>Проект "Комменты"</title>
-  <meta charset="utf-8" />
-  <link rel="stylesheet" href="styles.css" />
-</head>
-
-<body>
-  <div id="app"></div>
-  <script type="module" src="scripts/main.js"></script>
-</body>
-
+  <body>
+    <div id="app"></div>
+    <script src="./dist/main.js"></script>
+  </body>
 </html>
