+++ scripts/main.js
@@ -1,13 +1,15 @@
 "use strict";
 
+import { getFetchAndRender } from "./api.js";
+import { showLoadingIndicator } from "./helpers.js";
+import { addLoadingIndicator } from "../components/render-component.js";
 
-import { getFetchAndRender } from './api.js';
-
-// рендер имеющихся на сервере комментариев 
+// рендер имеющихся на сервере комментариев
 export let comments = [];
 export const setComments = (newComments) => {
-    comments = newComments;
-}
+  comments = newComments;
+};
 
+addLoadingIndicator();
+showLoadingIndicator();
 getFetchAndRender();
-
